Added: SkeletonBlendedAnimation.  Skeleton Blended Animations hold a collection of SkeletonAnimations and a blend amount for each.  When applied to a SkeletonBlendedGeometry, the verticies are blended by the defined amount for each of the attach SkeletonAnimations.
Added: SkeletonBlendedGeometry.  Skeleton Blended Geometry inherits off
of Geometry.  It adds the ability to attach Joints of a skeleton, by
some blend amount, to the verticies of the geometry.  When the skeleton
changes, then the Skeleton Blended Geometry recalculates the positions
of it's vertices and normals.

Added: Skeleton Blended Mesh Tutorial.