Changed: ParticleSystemCore now uses radix sort when sorting particles.
Added: Particles have IDs associated with them when they are created.
This ID can be used to index the same particle, or remove it from the
system.  This cannot be done by using the index of a particle, because
the index is not garanteed to remain the same through a particles
lifespan.