Added: Some MultiField editing controls.  The editable values of MultiFields are placed in a List, where each item of the list is a FieldEditor attached to that specific index of the MultiField.
Added: A feature to List that determines whether items in the list are
selectable.  If the Selectable field in List is false then none of the
items can be selected, as far as the List selection is concerned.  It is
still possible to interact with the Labels, buttons, etc that make up
the List items.