Added: AttachColGeomGraphOp, this GraphOp can add TriMesh, Sphere, Box, Capsule Geoms to a SceneGraph by matching against node names of TravMasks.  Geoms can also be removed with this GraphOp.
Fixed: Most of the Physics FieldContainers did not define the correct group
for their type definition.  This causes problems with attaching multiple Physics
Attachments.  The correct group definitions have been updated in the fcd files.

Fixed: Added a commit changes call in the PhysicsHandler before the physics
system is updated.