Added: Support for Focused Border, Background, Foreground, and Text color in DefaultTreeComponentGenerator.
Added: Support for moving the caret to a specific line, getting the
start and end position of a specific line in TextArea.

Changed: Tree no longer takes into account whether the expand component
of an element has focus when generating the component for that element.

Changed: Increased the default line height of Tree elements for the
WindowsLookAndFeel.