Fixed: Graphics2D and UIFont now use scale of the TXFFace when scaling the drawing and bounds of a text layout.
Changed: By default UIFont has AntiAliasing set to false.  When AntiAliasing is false
for a UIFont, GL_NEAREST is used as the min and max texture filters.  When
AntiAliasing is true, GL_LINEAR_MIPMAP_NEAREST is used as the min filter and
GL_LINEAR as the max filter.

This change makes text clearer when turning AntiAliasing off for UserInterfaces that
are drawn on a foreground.  However, when the user interface is draw in the scene,
turning AntiAliasing on gives better results.